--- a/d/source/app.d
+++ b/d/source/app.d
@@ -21,7 +21,7 @@ struct RelatedPosts
 
 struct PostsWithSharedTags
 {
-	ulong post;
+	uint post;
 	ubyte sharedTags;
 }