Minor performance tweak