@Imperatorn 2023-10-18
Note, this is not using size_t or ulong, but uint.

On my system, this gains about 8%