@cyrusmsk 2023-10-18
I've tried the similar PR before, but to int, instead of uint. And we had a regression in speed https://github.com/jinyus/related_post_gen/pull/269
And on local system I also can see improvement with this change.

Dear @jinyus, could you please check performance one more time with uint before merging?