@jinyus 2023-10-18
Main branch:
```
24.87 ms
317.38 ms
2.71 s
```

This PR:
```
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 33.058ms
        Processing time (w/o IO): 33.007ms
        Processing time (w/o IO): 32.991ms
        Processing time (w/o IO): 33.054ms
        Processing time (w/o IO): 32.5ms
        Processing time (w/o IO): 32.843ms
        Processing time (w/o IO): 33.035ms
        Processing time (w/o IO): 33ms
        Processing time (w/o IO): 32.982ms
        Processing time (w/o IO): 32.902ms
        Processing time (w/o IO): 33.14ms
        Processing time (w/o IO): 33.024ms
        Processing time (w/o IO): 32.53ms
          Time (mean ± σ):     114.8 ms ±   0.8 ms    [User: 104.0 ms, System: 12.0 ms]
          Range (min … max):   113.7 ms … 116.3 ms    10 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 443.778ms
        Processing time (w/o IO): 443.212ms
        Processing time (w/o IO): 443.22ms
          Time (mean ± σ):     745.5 ms ±   1.7 ms    [User: 718.2 ms, System: 31.5 ms]
          Range (min … max):   744.3 ms … 746.7 ms    2 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 3829.38ms
        Processing time (w/o IO): 3828.65ms
        Processing time (w/o IO): 3834.43ms
          Time (mean ± σ):      4.792 s ±  0.011 s    [User: 4.675 s, System: 0.123 s]
          Range (min … max):    4.785 s …  4.799 s    2 runs

```