@jinyus 2023-10-18
Fresh run on main:
```
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 25.426ms
        Processing time (w/o IO): 24.884ms
        Processing time (w/o IO): 24.711ms
        Processing time (w/o IO): 24.354ms
        Processing time (w/o IO): 25.259ms
        Processing time (w/o IO): 24.784ms
        Processing time (w/o IO): 24.87ms
        Processing time (w/o IO): 24.132ms
        Processing time (w/o IO): 25.003ms
        Processing time (w/o IO): 25.197ms
        Processing time (w/o IO): 24.683ms
        Processing time (w/o IO): 24.897ms
        Processing time (w/o IO): 24.693ms
          Time (mean ± σ):     105.0 ms ±   1.3 ms    [User: 97.1 ms, System: 9.3 ms]
          Range (min … max):   103.5 ms … 108.4 ms    10 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 316.677ms
        Processing time (w/o IO): 316.934ms
        Processing time (w/o IO): 315.443ms
          Time (mean ± σ):     611.4 ms ±   2.2 ms    [User: 588.7 ms, System: 27.2 ms]
          Range (min … max):   609.9 ms … 613.0 ms    2 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 2701.67ms
        Processing time (w/o IO): 2707.95ms
        Processing time (w/o IO): 2706.85ms
          Time (mean ± σ):      3.648 s ±  0.017 s    [User: 3.547 s, System: 0.105 s]
          Range (min … max):    3.636 s …  3.660 s    2 runs      
```